+++ pulse2percept/topography/retina/__init__.py
@@ -5,15 +5,18 @@
 
     base
     curcio1990
+    montesano2020
     watson2014
 
 """
 from .base import RetinalMap
 from .curcio1990 import Curcio1990Map
+from .montesano2020 import Montesano2020Map
 from .watson2014 import Watson2014Map, Watson2014DisplaceMap
 
 __all__ = [
     'Curcio1990Map',
+    'Montesano2020Map',
     'RetinalMap',
     'Watson2014DisplaceMap',
     'Watson2014Map',